Add instructor label to courses
Right now, Grapefruit doesn't support multiple instructors for one
course. And it should!

To put a temporary band-aid on the situation for now, we are putting in
an "instructor label" for the course that can indicate the presence of
multiple instructors

A long term solution should fully support multiple instructors in a
course.